requirements.txt
accelerate==0.30.0
colour==0.1.5
diffusers==0.27.2
einops==0.8.0
lpips==0.1.4
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.9.0.80
packaging==24.1
pandas==2.0.3
peft==0.10.0
Pillow==10.3.0
Requests==2.32.3
scipy==1.13.1
sd3d.egg==info
setuptools==69.5.1
skimage==0.0
torch==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.66.4
transformers==4.36.2
ultralytics==8.2.27
wandb==0.17.1
xformers==0.0.26.post1
